web-easy-broken_reset_password/deploy/app/requirements.txt
fastapi
uvicorn
pydantic
sqlmodel
PyJWT
python-dotenv
jinja2